CP-6096: Move existing iSCSI code to a base class
As a temporary step, move the existing code to a new base class
and remove the ISCSISR type.

In HBASR.py the import from ISCSISR has just been removed because
it is not needed.
